Added season e2e test
